change monitor

+++ utils.c
@@ -6,17 +6,18 @@
 /*   By: cvenkman <cvenkman@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/18 18:23:12 by cvenkman          #+#    #+#             */
-/*   Updated: 2021/10/23 21:35:25 by cvenkman         ###   ########.fr       */
+/*   Updated: 2021/10/24 03:01:12 by cvenkman         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-unsigned int	get_time(void)
+unsigned long long	get_time(void)
 {
 	struct timeval	time;
 
 	gettimeofday(&time, NULL);
+	// printf("+++  %d\n", (time.tv_sec * 1000) + (time.tv_usec / 1000));
 	return ((time.tv_sec * 1000) + (time.tv_usec / 1000));
 }
